# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API description for [Moira Alert API](https://moira.readthedocs.io/en/latest/overview.html).",
        "title": "Moira Alert",
        "contact": {
            "name": "Contact Moira Team",
            "email": "opensource@skbkontur.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "master"
    },
    "basePath": "/api",
    "paths": {
...
        "/event/all": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Deletes all notification events",
                "operationId": "delete-all-events",
                "responses": {
                    "200": {
                        "description": "Events removed successfully"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
        "/event/{triggerID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Gets all trigger events for current page and their count",
                "operationId": "get-events-list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "The ID of updated trigger",
                        "name": "triggerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "Number of items to be displayed on one page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Defines the number of the displayed page. E.g, p=2 would display the 2nd page",
                        "name": "p",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.EventsList"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorNotFoundExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
...
        "/notification": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Gets a paginated list of notifications, all notifications are fetched if end = -1 and start = 0",
                "operationId": "get-notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Default Value: 0",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "Default Value: -1",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.NotificationsList"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Delete a notification by id",
                "operationId": "delete-notification",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "The ID of updated trigger",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification have been deleted",
                        "schema": {
                            "$ref": "#/definitions/dto.NotificationDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
        "/notification/all": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Deletes all available notifications",
                "operationId": "delete-all-notifications",
                "responses": {
                    "200": {
                        "description": "All notifications have been deleted"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
        "/pattern": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pattern"
                ],
                "summary": "Get all patterns",
                "operationId": "get-all-patterns",
                "responses": {
                    "200": {
                        "description": "Patterns fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.PatternList"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
        "/pattern/{pattern}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pattern"
                ],
                "summary": "Deletes a Moira pattern",
                "operationId": "delete-pattern",
                "parameters": [
                    {
                        "type": "string",
                        "default": "DevOps.my_server.hdd.freespace_mbytes",
                        "description": "Trigger pattern to operate on",
                        "name": "pattern",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pattern deleted successfully"
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorInternalServerExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "server error during request handling"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "api.ErrorInvalidRequestExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource with the ID does not exist"
                },
                "status": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "api.ErrorNotFoundExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource with ID '66741a8c-c2ba-4357-a2c9-ee78e0e7' does not exist"
                },
                "status": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
...
        "api.ErrorRenderExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "rendering error"
                },
                "status": {
                    "type": "string",
                    "example": "Error rendering response"
                }
            }
        },
...
        "dto.EventsList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moira.NotificationEvent"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 0
                },
                "size": {
                    "type": "integer",
                    "example": 100
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "dto.NotificationDeleteResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "dto.NotificationsList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moira.ScheduledNotification"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "dto.PatternData": {
            "type": "object",
            "properties": {
                "metrics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "DevOps.my_server.hdd.freespace_mbytes",
                        " DevOps.my_server.hdd.freespace_mbytes",
                        " DevOps.my_server.db.*"
                    ]
                },
                "pattern": {
                    "type": "string",
                    "example": "Devops.my_server.*"
                },
                "triggers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TriggerModel"
                    }
                }
            }
        },
        "dto.PatternList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PatternData"
                    }
                }
            }
        },
...
        "dto.TriggerModel": {
            "type": "object",
            "properties": {
                "alone_metrics": {
                    "description": "A list of targets that have only alone metrics",
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    },
                    "example": {
                        "t1": true
                    }
                },
                "created_at": {
                    "description": "Datetime when the trigger was created",
                    "type": "string"
                },
                "created_by": {
                    "description": "Username who created trigger",
                    "type": "string"
                },
                "desc": {
                    "description": "Description string",
                    "type": "string",
                    "example": "check the size of /var/log"
                },
                "error_value": {
                    "description": "ERROR threshold",
                    "type": "number",
                    "example": 1000
                },
                "expression": {
                    "description": "Used if you need more complex logic than provided by WARN/ERROR values",
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "description": "Trigger unique ID",
                    "type": "string",
                    "example": "292516ed-4924-4154-a62c-ebe312431fce"
                },
                "is_remote": {
                    "description": "Shows if trigger is remote (graphite-backend) based or stored inside Moira-Redis DB",
                    "type": "boolean",
                    "example": false
                },
                "mute_new_metrics": {
                    "description": "If true, first event NODATA → OK will be omitted",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "Trigger name",
                    "type": "string",
                    "example": "Not enough disk space left"
                },
                "patterns": {
                    "description": "Graphite patterns for trigger",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        ""
                    ]
                },
                "sched": {
                    "description": "Determines when Moira should monitor trigger",
                    "allOf": [
                        {
                            "$ref": "#/definitions/moira.ScheduleData"
                        }
                    ]
                },
                "tags": {
                    "description": "Set of tags to manipulate subscriptions",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "server",
                        "disk"
                    ]
                },
                "targets": {
                    "description": "Graphite-like targets: t1, t2, ...",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "devOps.my_server.hdd.freespace_mbytes"
                    ]
                },
                "trigger_type": {
                    "description": "Could be: rising, falling, expression",
                    "type": "string",
                    "example": "rising"
                },
                "ttl": {
                    "description": "When there are no metrics for trigger, Moira will switch metric to TTLState state after TTL seconds",
                    "type": "integer",
                    "example": 600
                },
                "ttl_state": {
                    "description": "When there are no metrics for trigger, Moira will switch metric to TTLState state after TTL seconds",
                    "type": "string",
                    "example": "NODATA"
                },
                "updated_at": {
                    "description": "Datetime  when the trigger was updated",
                    "type": "string"
                },
                "updated_by": {
                    "description": "Username who updated trigger",
                    "type": "string"
                },
                "warn_value": {
                    "description": "WARN threshold",
                    "type": "number",
                    "example": 500
                }
            }
        },
...
        "moira.ContactData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1dd38765-c5be-418d-81fa-7a5f879c2315"
                },
                "team": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "mail"
                },
                "user": {
                    "type": "string",
                    "example": ""
                },
                "value": {
                    "type": "string",
                    "example": "devops@example.com"
                }
            }
        },
        "moira.EventInfo": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "integer",
                    "example": 0
                },
                "maintenance": {
                    "$ref": "#/definitions/moira.MaintenanceInfo"
                }
            }
        },
        "moira.MaintenanceInfo": {
            "type": "object",
            "properties": {
                "remove_time": {
                    "type": "integer",
                    "example": 0
                },
                "remove_user": {
                    "type": "string"
                },
                "setup_time": {
                    "type": "integer",
                    "example": 0
                },
                "setup_user": {
                    "type": "string"
                }
            }
        },
...
        "moira.NotificationEvent": {
            "type": "object",
            "properties": {
                "contactId": {
                    "type": "string"
                },
                "event_message": {
                    "$ref": "#/definitions/moira.EventInfo"
                },
                "metric": {
                    "type": "string",
                    "example": "carbon.agents.*.metricsReceived"
                },
                "msg": {
                    "type": "string"
                },
                "old_state": {
                    "type": "string",
                    "example": "ERROR"
                },
                "state": {
                    "type": "string",
                    "example": "OK"
                },
                "sub_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1590741878
                },
                "trigger_event": {
                    "type": "boolean",
                    "example": true
                },
                "trigger_id": {
                    "type": "string",
                    "example": "5ff37996-8927-4cab-8987-970e80d8e0a8"
                },
                "value": {
                    "type": "number",
                    "example": 70
                },
                "values": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                }
            }
        },
        "moira.PlottingData": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "theme": {
                    "type": "string",
                    "example": "dark"
                }
            }
        },
        "moira.ScheduleData": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moira.ScheduleDataDay"
                    }
                },
                "endOffset": {
                    "type": "integer",
                    "example": 1439
                },
                "startOffset": {
                    "type": "integer",
                    "example": 0
                },
                "tzOffset": {
                    "type": "integer",
                    "example": -60
                }
            }
        },
        "moira.ScheduleDataDay": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Mon"
                }
            }
        },
        "moira.ScheduledNotification": {
            "type": "object",
            "properties": {
                "contact": {
                    "$ref": "#/definitions/moira.ContactData"
                },
                "event": {
                    "$ref": "#/definitions/moira.NotificationEvent"
                },
                "plotting": {
                    "$ref": "#/definitions/moira.PlottingData"
                },
                "send_fail": {
                    "type": "integer",
                    "example": 0
                },
                "throttled": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1594471927
                },
                "trigger": {
                    "$ref": "#/definitions/moira.TriggerData"
                }
            }
        },
...
        "moira.TriggerData": {
            "type": "object",
            "properties": {
                "__notifier_trigger_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "server",
                        "disk"
                    ]
                },
                "desc": {
                    "type": "string",
                    "example": "check the size of /var/log"
                },
                "error_value": {
                    "type": "number",
                    "example": 1000
                },
                "id": {
                    "type": "string",
                    "example": "292516ed-4924-4154-a62c-ebe312431fce"
                },
                "is_remote": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "Not enough disk space left"
                },
                "targets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "devOps.my_server.hdd.freespace_mbytes"
                    ]
                },
                "warn_value": {
                    "type": "number",
                    "example": 5000
                }
            }
        }
    },
...
}